requirements.txt
pdf2docx>=0.5.6
python-docx>=0.8.11
lxml>=4.9.0
Pillow>=9.0.0
